Reworked polyfill support for Zend\Stdlib\ArrayObject
- Instead of using an autoload file, instead created two separate classes, named
  after the type of support they provide. The file Zend\Stdlib\ArrayObject now
  simply has a conditional in it; based on the condition met, a class_alias() is
  created -- and from then on, that alias will be used for
  Zend\Stdlib\ArrayObject.

- Removed autoload.files entries from composer.json files; no longer necessary.

- One test failure currently in doing a comparison of merged array objects; will
  investigate in a later commit.
