+++ library/Zend/Session/composer.json
@@ -9,10 +9,7 @@
     "autoload": {
         "psr-0": {
             "Zend\\Session\\": ""
-        },
-        "files": [
-            "Zend/Session/compatibility/autoload.php"
-        ]
+        }
     },
     "target-dir": "Zend/Session",
     "require": {
